server/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["vikyw89 <vikyw89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.10"
fastapi = "^0.110.2"
prisma = "^0.13.1"
llama-index = "^0.10.31"
langgraph = "^0.0.38"
nest-asyncio = "^1.6.0"
python-dotenv = "^1.0.1"
langchain-community = "^0.0.34"
google-cloud-aiplatform = "^1.48.0"
llama-index-llms-litellm = "^0.1.4"
langchain-google-genai = "^1.0.2"
redis = "^5.0.4"
jsonpickle = "^3.0.4"
numpy = "^1.26.4"
pinecone-client = "^3.2.2"
llama-index-vector-stores-pinecone = "^0.1.6"
llama-index-postprocessor-flag-embedding-reranker = "^0.1.2"
streamlit = "^1.33.0"
langsmith = "^0.1.52"
unsync = "^1.4.0"
llama-index-graph-stores-neo4j = "^0.1.4"
llama-index-llms-openai = "^0.1.16"
audio-recorder-streamlit = "^0.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "scripts.start:run"
testing = "scripts.testing:run"
seed_vectordb = "scripts.seed_vectordb:run"